+++ packages/alurkerja-ui/src/components/TableLowcode/TableLowcode.stories.tsx
@@ -81,6 +81,7 @@ export const BPMN: Story = {
         setFilterBy={setFilterBy}
         search={search}
         setSearch={setSearch}
+        column={[{ key: 'alasan_cuti', label: 'Halo' }]}
       />
     )
   },