+++ packages/alurkerja-ui/src/hooks/useFieldOrder.ts
@@ -2,7 +2,7 @@ import { FieldProperties } from '@/types'
 import { useEffect, useState } from 'react'
 
 export const useFieldOrder = ({ fields }: { fields?: { [x: string]: FieldProperties } }) => {
-  const [listFieldKey, setFieldkey] = useState<string[]>()
+  const [listFieldKey, setFieldkey] = useState<{ label: string; key: string }[]>()
 
   useEffect(() => {
     if (fields) {
@@ -10,7 +10,7 @@ export const useFieldOrder = ({ fields }: { fields?: { [x: string]: FieldPropert
         Object.entries(fields)
           .map(([_key, value]) => value)
           .sort((a, b) => a.list_order - b.list_order)
-          .map((field) => field.name)
+          .map((field) => ({ label: field.label, key: field.name }))
       )
     }
   }, [fields])